pyproject.toml
[tool.poetry]
name = "kelishamiz-backend"
version = "0.1.0"
description = ""
authors = ["murtazoxurramov <murtazoxurramov1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
eskiz-sms = "^0.2.3"
django-phonenumbers = "^1.0.1"
python-decouple = "^3.8"
pillow = "^10.0.0"
gunicorn = "^21.2.0"
django-filter = "^23.3"
django-cors-headers = "^4.3.0"
django-ninja-jwt = "^5.2.10"
django-ninja = "^1.1.0"
django-ninja-extra = "^0.20.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
